Refactor proxy service interface to reduce package cycles.
