+++ project-set/core/valve/src/main/java/com/rackspace/cloud/valve/http/proxy/jerseyclient/JerseyClientProxyService.java
@@ -17,10 +17,11 @@
 package com.rackspace.cloud.valve.http.proxy.jerseyclient;
 
 import java.net.MalformedURLException;
+
+import com.rackspace.cloud.valve.http.proxy.common.ProxyService;
 import org.apache.http.HttpException;
 import org.slf4j.LoggerFactory;
 import org.slf4j.Logger;
-import com.rackspace.cloud.valve.http.proxy.ProxyService;
 import com.rackspace.cloud.valve.http.proxy.httpclient.HttpResponseCodeProcessor;
 import com.sun.jersey.api.client.Client;
 import com.sun.jersey.api.client.ClientResponse;
